Refactored backend to live in shared library.
