+++ gui-js/apps/minsky-electron/src/app/managers/BookmarkManager.ts
@@ -1,9 +1,8 @@
-import { DescriptionPayload } from '@minsky/shared';
+import { DescriptionPayload, minsky } from '@minsky/shared';
 import { Menu, MenuItem } from 'electron';
 import { CommandsManager } from './CommandsManager';
 import { callRESTApi } from './RestServiceManager';
 import { WindowManager } from './WindowManager';
-import {minsky} from '../backend';
 const JSON5 = require('json5');
