+++ gui-js/apps/minsky-electron/src/app/managers/WindowManager.ts
@@ -7,6 +7,7 @@ import {
   OPEN_DEV_TOOLS_IN_DEV_BUILD,
   rendererAppName,
   rendererAppURL,
+  Utility,
 } from '@minsky/shared';
 //import * as debug from 'debug';
 import { BrowserWindow, dialog, Menu, screen } from 'electron';
@@ -14,7 +15,6 @@ import * as log from 'electron-log';
 import * as os from 'os';
 import { join } from 'path';
 import { format } from 'url';
-import { Utility } from '../utility';
 
 //const logWindows = debug('minsky:electron_windows');
