+++ gui-js/apps/minsky-web/src/main.ts
@@ -9,6 +9,7 @@ if (AppConfig.production) {
   enableProdMode();
 }
 
+
 platformBrowserDynamic()
   .bootstrapModule(AppModule, {
     preserveWhitespaces: false,
