+++ gui-js/libs/shared/src/index.ts
@@ -1,3 +1,5 @@
+export * from './lib/backend/index';
+export * from './lib/environment';
 export * from './lib/constants/constants';
 export * from './lib/constants/magicNumbers';
 export * from './lib/constants/MainRenderingTabs';
@@ -7,3 +9,4 @@ export * from './lib/interfaces/ActiveWindow';
 export * from './lib/interfaces/Interfaces';
 export * from './lib/material/material.module';
 export * from './lib/shared.module';
+export * from './lib/utility';
