+++ gui-js/libs/shared/src/lib/backend/canvas.ts
