+++ gui-js/libs/shared/src/lib/backend/godleyIcon.ts
