+++ gui-js/libs/shared/src/lib/backend/operation.ts
