+++ gui-js/libs/shared/src/lib/backend/plot.ts
