+++ gui-js/libs/shared/src/lib/backend/plotTab.ts
