+++ gui-js/libs/shared/src/lib/backend/renderNativeWindow.ts
