+++ gui-js/libs/shared/src/lib/backend/variable.ts
