+++ gui-js/libs/shared/src/lib/environment.ts
