+++ gui-js/libs/shared/src/lib/utility.ts
@@ -1,7 +1,7 @@
-import { environment } from '../environments/environment';
+import { environment } from './environment';
 
 export class Utility {
-  public static isDevelopmentMode() {
+  public static isDevelopmentMode(): boolean {
     const isEnvironmentSet: boolean = 'ELECTRON_IS_DEV' in process.env;
     const getFromEnvironment: boolean =
       parseInt(process.env.ELECTRON_IS_DEV, 10) === 1;
@@ -9,7 +9,7 @@ export class Utility {
     return isEnvironmentSet ? getFromEnvironment : !environment.production;
   }
 
-  public static isPackaged() {
+  public static isPackaged(): boolean {
     // https://github.com/ganeshrvel/npm-electron-is-packaged/blob/master/lib/index.js
 
     let isPackaged = false;
@@ -27,4 +27,9 @@ export class Utility {
 
     return isPackaged;
   }
+
+//  // @return true if this is the renderer process
+//  public static isRenderer(): boolean {
+//    return typeof window!=='undefined' && typeof window.process==='object' && window.process?.type==='renderer';
+//  }
 }
