+++ gui-js/libs/shared/src/test-setup.ts
@@ -1,1 +1,2 @@
 import 'jest-preset-angular/setup-jest';
+import '../../../apps/minsky-electron/src/app/backend-init';
